Traits&HoverOver
TRaits can now simply be made as a new cardButton in the located folder, which will then be added to the list of traits.

Traits and abilities can also be hovered over with the mouse, which displays the correalted card giving the player information of what skill/trait it is.

Also added some new skills and traits as well as their respective icon/sprite.