Inventory_Implemented
Implemented the inventory and created a card system for gear which auto adjust.
Updated some materials.
Updated som abilities and traits texts and functions.
Added functions to more easily add and remove all of the misc stats.
Above mentioned stats also are connected to a stats panel which are updated.